requirements.txt
boto3
botocore
click
docutils
Flask
Flask-SQLAlchemy
flask-redis
gevent
greenlet
gunicorn
itsdangerous
Jinja2
jmespath
MarkupSafe
PyMySQL
python-dateutil
python-dotenv
s3transfer
sacrud
six
SQLAlchemy
sqlalchemy-mixins
typing
watchtower
Werkzeug